Update manipulating_the_hangman_db.py
This is a separate file of all the SQL commands used in the main game file. This file doesn't need to be run in conjunction with the game file, however I thought it would useful to include on this repository for two reasons: 1. to group all SQL commands used during game into one file that can be quickly read and understood by others 2. it is useful to have a separate file for developing features that incorporate SQL.